+++ Runtime/Ponito/Core/Asyncs/Tasks/Movables/DelayAwait.cs
@@ -1,4 +1,5 @@
 ﻿using System.Threading;
+using Ponito.Core.Asyncs.Interfaces;
 using UnityEngine;
 
 namespace Ponito.Core.Asyncs.Tasks.Movables
@@ -6,7 +7,7 @@ namespace Ponito.Core.Asyncs.Tasks.Movables
     /// <summary>
     ///     Delays a certain amount of time.
     /// </summary>
-    internal class DelayAwait : MovableBase
+    internal class DelayAwait : MovableBase, ProgressMovable
     {
         /// <summary>
         ///     By <see cref="milliseconds" />
@@ -14,7 +15,8 @@ namespace Ponito.Core.Asyncs.Tasks.Movables
         public DelayAwait(int milliseconds, CancellationToken ct = default)
         {
             Ct         = ct;
-            targetTime = Time.time + milliseconds / 1000f;
+            seconds    = milliseconds / 1000f;
+            targetTime = Time.time + seconds;
         }
 
         /// <summary>
@@ -22,14 +24,20 @@ namespace Ponito.Core.Asyncs.Tasks.Movables
         /// </summary>
         public DelayAwait(float seconds, CancellationToken ct = default)
         {
-            Ct         = ct;
-            targetTime = Time.time + seconds;
+            Ct           = ct;
+            this.seconds = seconds;
+            targetTime   = Time.time + this.seconds;
         }
 
         /// <summary>
-        ///     Target <see cref="Time" />.<see cref="Time.time" />
+        ///     目標 <see cref="Time" />.<see cref="Time.time" />
         /// </summary>
         private float targetTime { get; }
+
+        /// <summary>
+        ///     要等候的秒數
+        /// </summary>
+        private float seconds { get; }
 
         /// <inheritdoc />
         public override bool MoveNext()
@@ -42,5 +50,10 @@ namespace Ponito.Core.Asyncs.Tasks.Movables
 
             return ContinueMoveNext();
         }
+
+        /// <summary>
+        ///     回報等候的進度
+        /// </summary>
+        public float Progress => 1f - (targetTime - Time.time) / seconds;
     }
 }