+++ Runtime/Ponito/Core/Asyncs/Tasks/PoTaskBase.cs
@@ -1,9 +1,6 @@
 using System;
 using System.Runtime.CompilerServices;
 using Ponito.Core.Asyncs.Interfaces;
-using Ponito.Core.Asyncs.Tasks.Movables;
-using Ponito.Core.Asyncs.Extensions;
-using UnityEngine;
 
 namespace Ponito.Core.Asyncs.Tasks
 {
@@ -12,8 +9,6 @@ namespace Ponito.Core.Asyncs.Tasks
     /// </summary>
     public abstract class PoTaskBase
     {
-        internal object name { get; set; }
-        
         /// <summary>
         ///     儲存內部 <see cref="Movable"/> 
         /// </summary>